requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
exifread>=3.0.0
